requirements.txt
eventsourcing[sqlalchemy]
fastapi
starlette-context
pytest-asyncio
ariadne
uvicorn
